Правки по замечаниям.
