+++ package.json
@@ -38,6 +38,7 @@
     "react-dom": "16.12.0",
     "react-redux": "7.2.0",
     "react-router-dom": "5.1.2",
-    "redux": "4.0.5"
+    "redux": "4.0.5",
+    "redux-act": "1.7.7"
   }
 }
