+++ src/redux/reducer.js
@@ -1,0 +1,6 @@
+import {combineReducers} from 'redux';
+import {reducer as catalogReducer} from './catalog/reducer';
+
+export default combineReducers({
+  catalog: catalogReducer,
+});
